feat(core): Move typings to root directory
This commit moves the `typings` directory to the root of the project.
This change affects the import paths in `runStackCommand.ts` and `logger.ts`, which have been updated accordingly.

The diff includes:
- Update import path for `postToClient` in `runStackCommand.ts` to `~/handlers/modules/live-stacks`.
- Update import path for `Stack` in `runStackCommand.ts` to `~/../typings/docker-compose`.
- Update import path for `log_message` in `logger.ts` to `../../../typings/database`.
- Moves the typings dir to the root of the project

The reason for this change is to simplify the project structure and make the typings more accessible to other parts of the application.