✨ Feat: Add theme options
This commit introduces the ability to store and retrieve theme options in the database.

The `themes` table in the database has been updated to include an `options` column, which stores the theme options as a JSON string. The `addTheme` function now accepts an `options` parameter and stores it in the database. A new function `getThemeOptions` has been added to retrieve the theme options for a given theme.

This change allows for more flexible and customizable themes.